Move OpDef into resource.rs
