+++ src/ops.rs
@@ -12,7 +12,7 @@ pub mod validate;
 use crate::types::{EdgeKind, Signature, SignatureDescription};
 use crate::{Direction, Port};
 
-pub use custom::{CustomOp, OpDef, OpaqueOp};
+pub use custom::{CustomOp, OpaqueOp};
 
 use portgraph::NodeIndex;
 use smol_str::SmolStr;
